Move do_if rules to pipeline package
