+++ pipeline/plugin.go
@@ -5,7 +5,6 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/ozontech/file.d/cfg/matchrule"
 	"github.com/ozontech/file.d/metric"
 	"go.uber.org/zap"
 )
@@ -98,7 +97,7 @@ type ActionPluginStaticInfo struct {
 	MatchMode        MatchMode
 	MatchInvert      bool
 
-	DoIfChecker *matchrule.DoIfChecker
+	DoIfChecker *DoIfChecker
 }
 
 type ActionPluginInfo struct {
